refactor: dead code removal, dedup export logic, extract SpeciesHierarchyBuilder
- Remove FocusPointStatus enum, bestEyeVisibility parameter (dead since pipeline refactor)
- Remove processingFilename (written but never read)
- Remove stale Skill Level localization strings
- Move SidebarSelection enum from SourceListView to AppState (it's a model type)
- Extract performExport() helper in MainView, dedup exportPicks/exportAllVisible
- Extract SpeciesHierarchyBuilder.build(from:) as pure static function from AppState
