+++ apps/mac-client/SuperPickyApp/AppState.swift
@@ -1,6 +1,16 @@
 import SwiftUI
 import os
 
+enum SidebarSelection: Hashable {
+    case folder(URL)
+    case rating(Int)
+    case flying
+    case picks
+    case species(String)
+    case burstGroup(UUID)
+    case singles(String) // species name
+}
+
 /// Species with its burst groups for the sidebar hierarchy.
 struct SpeciesEntry: Identifiable {
     let name: String
@@ -52,7 +62,6 @@ final class AppState {
     // Per-folder processing state
     var processingFolder: URL?
     var processingProgress: Double = 0
-    var processingFilename: String = ""
     var currentFolder: URL?
 
     var isProcessing: Bool { processingFolder != nil }
@@ -113,79 +122,7 @@ final class AppState {
 
     /// Build species → burst group hierarchy from loaded photos.
     private func buildSpeciesHierarchy() {
-        // Assign each burst to its dominant species by highest confidence ID.
-        // A burst spanning multiple species classifications appears only once.
-        var burstToSpecies: [UUID: String] = [:]
-        var burstBestConfidence: [UUID: (species: String, confidence: Float)] = [:]
-        var burstPhotos: [UUID: [Photo]] = [:]
-
-        for photo in allPhotos {
-            guard let groupID = photo.burstGroupID else { continue }
-            burstPhotos[groupID, default: []].append(photo)
-
-            guard let name = photo.speciesCommonName ?? photo.speciesScientificName else { continue }
-            let confidence = photo.speciesConfidence ?? 0
-            if let current = burstBestConfidence[groupID] {
-                if confidence > current.confidence {
-                    burstBestConfidence[groupID] = (name, confidence)
-                }
-            } else {
-                burstBestConfidence[groupID] = (name, confidence)
-            }
-        }
-
-        for groupID in burstPhotos.keys {
-            burstToSpecies[groupID] = burstBestConfidence[groupID]?.species ?? "Unidentified"
-        }
-
-        // Group photos by species
-        var bySpecies: [String: (photos: [Photo], isUnidentified: Bool)] = [:]
-        for photo in allPhotos {
-            let hasSpecies = photo.speciesScientificName != nil
-            let name = photo.speciesCommonName ?? photo.speciesScientificName ?? String(localized: "Unidentified")
-            var entry = bySpecies[name] ?? (photos: [], isUnidentified: !hasSpecies)
-            entry.photos.append(photo)
-            bySpecies[name] = entry
-        }
-
-        speciesEntries = bySpecies.map { name, entry in
-            // Only include burst groups whose dominant species matches this entry
-            var burstGroupIDs: Set<UUID> = []
-            var singleCount = 0
-            for photo in entry.photos {
-                if let groupID = photo.burstGroupID {
-                    if burstToSpecies[groupID] == name {
-                        burstGroupIDs.insert(groupID)
-                    }
-                    // Photos in bursts owned by another species don't count as singles
-                } else {
-                    singleCount += 1
-                }
-            }
-
-            let burstGroups = burstGroupIDs.map { groupID in
-                let groupPhotos = burstPhotos[groupID] ?? []
-                let best = groupPhotos.first { $0.isBurstBest }
-                return BurstGroupEntry(
-                    id: groupID,
-                    count: groupPhotos.count,
-                    bestFilename: best?.filename ?? groupPhotos.first?.filename
-                )
-            }.sorted { $0.count > $1.count }
-
-            return SpeciesEntry(
-                name: name,
-                cnName: entry.photos.first?.speciesCnName,
-                count: entry.photos.count,
-                burstGroups: burstGroups,
-                singlePhotos: singleCount,
-                isUnidentified: entry.isUnidentified
-            )
-        }.sorted {
-            // Unidentified always first, then by count
-            if $0.isUnidentified != $1.isUnidentified { return $0.isUnidentified }
-            return $0.count > $1.count
-        }
+        speciesEntries = SpeciesHierarchyBuilder.build(from: allPhotos)
     }
 
     /// Clear all photo data (when folder is removed).
