+++ apps/mac-client/SuperPickyApp/PipelineCoordinator.swift
@@ -216,7 +216,6 @@ final class PipelineCoordinator {
             sharpness: photo.sharpnessScore ?? 0,
             aesthetics: photo.aestheticsScore,
             allKeypointsHidden: keypoints.allKeypointsHidden,
-            bestEyeVisibility: keypoints.bestEyeVisibility,
             isOverexposed: photo.exposureStatus == ExposureStatus.overexposed.rawValue,
             isUnderexposed: photo.exposureStatus == ExposureStatus.underexposed.rawValue,
             isFlying: flight.isFlying,
