+++ apps/mac-client/SuperPickyApp/en.lproj/Localizable.strings
@@ -103,8 +103,6 @@
 "Showing burst best only — click to show all" = "Showing burst best only — click to show all";
 "Showing picks only — click to show all" = "Showing picks only — click to show all";
 "Singles" = "Singles";
-"Skill Level" = "Skill Level";
-"Skill level sets default thresholds for sharpness and aesthetics scoring." = "Skill level sets default thresholds for sharpness and aesthetics scoring.";
 "Species" = "Species";
 "Star rating" = "Star rating";
 "System" = "System";
