+++ apps/mac-client/SuperPickyApp/zh-Hans.lproj/Localizable.strings
@@ -103,8 +103,6 @@
 "Showing burst best only — click to show all" = "仅显示连拍最佳 — 点击显示全部";
 "Showing picks only — click to show all" = "仅显示精选 — 点击显示全部";
 "Singles" = "单张";
-"Skill Level" = "技能等级";
-"Skill level sets default thresholds for sharpness and aesthetics scoring." = "技能等级决定了锐度和美学评分的默认阈值。";
 "Species" = "鸟种";
 "Star rating" = "星级评分";
 "System" = "跟随系统";
